Move configure function outside of start.tsx
